UGC-4220 | Add explicit replaceSelection() implementation
TemplateData interacts with the source editor via the $.textSelection()
wrapper, for which CodeEditor exposes a facade. This facade is missing
an explicit replaceSelection() implementation, in the absence of which
the plugin falls back to a naive implementation that errors out. So,
provide a simple implementation that calls the appropriate Ace method.